// src/app/models/scoring.model.ts
export class Scoring {
    idScoring: number;
    numeScoala: string;
    judetS: string;
    localitateS: string;
    numeU: string;
    statutU: string;
    romiUnitateP: number;
    dizabilitatiUnitateP: number;
    parintiUnitateP: number;
    burseUnitateP: number;
    totalCladire: number[];
    romiCladire: number[];
    dizabilitatiCladire: number[];
    parintiCladire: number[];
    burseCladire: number[];

    romiScorCladire: number[];
    dizabilitatiScorCladire: number[];
    parintiScorCladire: number[];
    burseScorCladire: number[];

    romiScorCladireIndex: number;
    dizabilitatiScorCladireIndex: number;
    parintiScorCladireIndex: number;
    burseScorCladireIndex: number;

    totalNivel: number[];
    romiNivel: number[];
    dizabilitatiNivel: number[];
    parintiNivel: number[];
    burseNivel: number[];
    repetentiNivel: number[];
    onlineNivel: number[];
    remedialaNivel: number[];

    romiScorNivel: number[];
    dizabilitatiScorNivel: number[];
    parintiScorNivel: number[];
    burseScorNivel: number[];
    repetentiScorNivel: number[];
    onlineScorNivel: number[];
    remedialaScorNivel: number[];

    romiScorNivelValue: number[];
    dizabilitatiScorNivelValue: number[];
    parintiScorNivelValue: number[];
    burseScorNivelValue: number[];
    repetentiScorNivelValue: number[];
    onlineScorNivelValue: number[];
    remedialaScorNivelValue: number[];

    romiScorNivelIndex: string[];
    dizabilitatiScorNivelIndex: string[];
    parintiScorNivelIndex: string[];
    burseScorNivelIndex: string[];
    repetentiScorNivelIndex: string[];
    onlineScorNivelIndex: string[];
    remedialaScorNivelIndex: string[];

    totalScor2b: number[];
    romiScor2b: number[];
    dizabilitatiScor2b: number[];
    parintiScor2b: number[];
    burseScor2b: number[];

    romiScor2bValue: number[];
    dizabilitatiScor2bValue: number[];
    parintiScor2bValue: number[];
    burseScor2bValue: number[];

    romiScor2bIndex: string[];
    dizabilitatiScor2bIndex: string[];
    parintiScor2bIndex: string[];
    burseScor2bIndex: string[];



    scorA1: number;
    scorA2: number;
    scorA3: number;
    scorA4: number;
    scorA5: number;
    scorB1: number;
    scorB2: number;
    scorB3: number;
    scorB4: number;
    scorB5: number;
    scorC1: number;
    scorC2: number;
    scorC3: number;
    scorC4: number;
    scorC5: number;
    scorC6: number;
    scorC7: number;
    scorC8: number;
    scorC9: number;
    scorC10: number;
    scorD1: number;
    scorD2: number;
    scorD3: number;
    scorE1: number;
    scorE2: number;
    scorE3: number;
    nrDep: string;
    nrDepC: string;
    scoalaId: number;
    tempWOW: number;

    uProcent: number;
    cProcent: number;

    sirues: string;
    mediu: string;
    mscolar: string;
    romani: string;
    finantari: string;
    plan: string;
    msanitar: string;
    expert: string;


    raspuns37: string;
    raspuns38: string;
    raspuns39: string;
    raspuns40: string;
    raspuns41: string;
    raspuns46: string;
    raspuns47: string;

    constructor(input?: any) {
        Object.assign(this, input);
    }
}